+++ src/background/initialization.ts
@@ -111,9 +111,7 @@ async function onInstalled({
     if (
       !(await storage.sync.get('hasInstructionsShown')).hasInstructionsShown
     ) {
-      openURL(
-        'https://github.com/crimx/ext-saladict/wiki/Instructions#wiki-content'
-      )
+      openURL('https://saladict.crimx.com/notice.html')
       storage.sync.set({ hasInstructionsShown: true })
     }
   } else if (reason === 'update') {